requirements.txt
nemo_toolkit[tts]
transformers @ git+https://github.com/huggingface/transformers.git
fastapi
uvicorn

